+++ packages/core/src/browser/cm-file-label-provider.ts
@@ -3,7 +3,7 @@
  ********************************************************************************/
 
 import { ModelService } from '@crossmodel/model-service/lib/common';
-import { ModelStructure } from '@crossmodel/protocol';
+import { ModelFileExtensions, ModelStructure } from '@crossmodel/protocol';
 import { Emitter, MaybePromise } from '@theia/core';
 import { DepthFirstTreeIterator, LabelProvider, LabelProviderContribution, Tree, TreeDecorator, TreeNode } from '@theia/core/lib/browser';
 import { WidgetDecoration } from '@theia/core/lib/browser/widget-decoration';
@@ -32,24 +32,10 @@ export class CrossModelLabelProvider implements LabelProviderContribution, TreeD
 
    getIcon(node: FileStatNode): string {
       if (this.isDataModelDirectory(node)) {
-         return ModelStructure.System.ICON_CLASS + ' default-folder-icon';
+         return ModelStructure.DataModel.ICON_CLASS + ' default-folder-icon';
       }
-      if (this.isDataModelDirectory(node.parent) && node.fileStat.name === ModelStructure.LogicalEntity.FOLDER) {
-         return ModelStructure.LogicalEntity.ICON_CLASS + ' default-folder-icon';
-      }
-      if (this.isDataModelDirectory(node.parent) && node.fileStat.name === ModelStructure.Relationship.FOLDER) {
-         return ModelStructure.Relationship.ICON_CLASS + ' default-folder-icon';
-      }
-      if (this.isDataModelDirectory(node.parent) && node.fileStat.name === ModelStructure.SystemDiagram.FOLDER) {
-         return ModelStructure.SystemDiagram.ICON_CLASS + ' default-folder-icon';
-      }
-      if (this.isDataModelDirectory(node.parent) && node.fileStat.name === ModelStructure.Mapping.FOLDER) {
-         return ModelStructure.Mapping.ICON_CLASS + ' default-folder-icon';
-      }
-      if (this.isDataModelDirectory(node.parent) && node.fileStat.name === ModelStructure.DataModel.FILE) {
-         return ModelStructure.DataModel.ICON_CLASS + ' default-file-icon';
-      }
-      return this.labelProvider.getIcon(node.fileStat);
+      const modelFileIcon = ModelFileExtensions.getIconClass(node.uri.toString());
+      return modelFileIcon ? modelFileIcon + ' default-file-icon' : this.labelProvider.getIcon(node.fileStat);
    }
 
    getName(node: FileStatNode): string | undefined {