+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0](https://github.com/crossmodel/crossmodel-core/compare/v0.7.5...v0.8.0) (2025-09-22)
+
+
+### Features
+
+* Add existing relationship to system diagram ([#171](https://github.com/crossmodel/crossmodel-core/issues/171)) ([d0bbdeb](https://github.com/crossmodel/crossmodel-core/commit/d0bbdeb9c70555f01abd0594993997e789e19abe))
+
+
+
+
+
 ## [0.7.5](https://github.com/crossmodel/crossmodel-core/compare/v0.7.4...v0.7.5) (2025-09-04)
 
 **Note:** Version bump only for package crossmodel-core